Remove the code related to puppet-glance from bare metal
The glance puppet module is not installed on bare metal, so the
system parameters and code that populates hiera data for those fields
is unused and can be removed.

This removes references to glance puppet operator from other
puppet operators.  Those other puppet operators are also deprecated and
are being removed by subsequent tasks.

Change-Id: I6c67804d83628039c5b93dcaf6f983830d156124
Story: 2004764
Task: 33767
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>