fix influxdb log file cannot be collected by logrotate
The logfile for influxdb is influxd.log, not influxdb.log
Correct the name in controller.yaml, so that logrotate could get
the log correctly.
Also, influxdb process uses error re-direction to append the error
message to influxd.log, so "copytruncate" need be added in logrotate
setting, so that log could be continue logged after logrotate
execution.

Closes-Bug: 1833167

Change-Id: I1ef484e3052355e5edc0cd0f7c629c7944daf014
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>