Add checks to deny adding secondary tiers before ceph is up
On controller-0, at initial configuration, adding secondary
tiers fails. The reason is that ceph has to be up in order
for secondary tiers and secondary storage backends to be
added. This operation is now denied through a semantic
checks.

Another semantic check is added to verify if ceph is
functional before allowing creation of secondary tiers.

Change-Id: I75c2e9540c98a01044bdd2666e0eac821f385cfa
Closes-Bug: 1830191
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>