Introduce ArmadaManifestOperator
Add the ArmadaManifestOperator to provide a single point of manipulation
for an application manifest.

This operator will load the original manifest file and provide a set of
utilily methods to update the list of chart_groups, an individual chart
group, or an individual chart based on platform and plugin conditions.
Based on the methods that are executed, changes are tracked and
overrides are generated for those elements that have changed.

These armada overrides are fed to the Armada command in addition to the
Helm overrides.

This also includes updated changes originally proposed from
I4bf4dcc3900c65f1492a362d31c9b3e7d73b8e31 to allow Barbican and
Telemetry chart groups to be optional services. This is utilizing a k8s
label to enable the services. This mechanism will be re-visited in a
future commit, but for now continue to use the existing mechanism.

Change-Id: I7e0c5bbe7daa0b0b8365ca1ed5f8fdf11a348c62
Partial-Bug: #1833746
Signed-off-by: Robert Church <robert.church@windriver.com>