Reraise exception and log error message
When the sysinv_api_workers is configured incorrect, the
sysinv api will not start due to the invalid default value.
therefore will cause host swact.

In this case, we can hardly figure out why that node is
failure, This patch is to catch exception and log error
message and then reraise exception if sysinv_api_workers
is invalid value.

Closes-bug: #1835142
Change-Id: Ie8dd185a6a24aec28c8834d5013cc003a1f7792f
Signed-off-by: junboli <junbo85.li@gmail.com>