Audit application status upon sysinv-conductor startup
If sysinv conductor process is abruptly terminated due to power
loss, OOM, missed audit cycle, etc... while an application upload/
apply/update/remove is in progres; the status of the application
is stuck in uploading/applying/updating/removing preventing any
subsequent system applications to resume (or revert) the operation.

This is the first commit of multi-commit change to improve the
robustness of the application framework.

Partial-Bug: 1833323
Change-Id: I73129d5621c77b50c2e29c078e6b99089244129f
Signed-off-by: Tee Ngo <Tee.Ngo@windriver.com>