Ensure deletion of conditional charts on application-remove
Armada delete does not support --value overrides files, so any
conditional manifest changes applied via overrides during apply will not
be reflected when the manifest is applied during delete.

With the addition of the ArmadaManifestOperator, a final view of the
manifest can be generated so that the armada delete has an accurate view
of what needs to be cleaned up.

Create this file an use it during the application-remove instead of the
original manifest.

Change-Id: I20191656805c5d8a419dbdeb7d2e97f87c5b2f2c
Depends-On: I7e0c5bbe7daa0b0b8365ca1ed5f8fdf11a348c62
Closes-Bug: #1833718
Signed-off-by: Robert Church <robert.church@windriver.com>