Retry application of managed apps after failure
platform-integ-apps as a managed app is automatically uploaded
and applied. It may happen that while the application is applied
the image registry it tries to download from is not available
so platform-integ-apps is left in apply-failed state with no
attempt to recover it. This requires manual intervention to
re-apply the app but can be automated as well.

Add two recovery mechanisms:

1. if one image download fails then restart the entire download
   procedure. Images that were already downloaded are be available in
   the local registry. The overhead of downloading them again is
   low.

2. automatically apply managed apps if they are in "apply-failed"
   caused by image download issues. Previously managed apps
   are automatically applied when in "uploaded" state. This
   happens after a larger timeout so the app can be manually
   managed (removed, etc.)

Promote nested k8s application audit functions to class scope
to increase readability and create auto-upload/apply/recover
managed app functions.

Closes-Bug: 1832032
Change-Id: Ic12b5ba89a28d34995444572d4c30ed583e5f7b4
Signed-off-by: Daniel Badea <daniel.badea@windriver.com>