Fix the huge pages fields showing as pending
The user expects to see the pending fields to be clear
after the huge pages modification and unlock is complete.
When sysinv-agent starts, it sends the host inventory update.
By this time, if the worker manifest apply has not been
completed, the memory report will be skipped. After the
initial report, the memory update will be triggered by
periodical audit. The audit runs every minute, but the
memory report is sent after 5 audit interval. As a result,
the huge pages settings could still show as pending on
CLI/GUI after around 10 minutes (more noticeable on AIO),
and this led the user to believe something went wrong with
the huge pages allocation (although the huge pages have
been allocated in Linux after the manifest is applied).

After talking with John Kung, we decided to send the memory
report in normal audit interval (every other minute).

Closes-Bug: 1830549

Change-Id: Idf5067648031168078d99a5d84c8368cbd400508
Signed-off-by: Tao Liu <tao.liu@windriver.com>