Add checks for dns name in certificate
This commit adds validation of https certificate for openstack
to ensure the provisioned endpoint domain matches the dns names
or the common name in the certificate.

If the domain is example.com, then the CN or SAN value must be
*.example.com.
SAN values, such as *.*.example.com, bad.*.example.com would not be
accepted.

Closes-Bug: 1826227

Change-Id: Ib50491282e8bf15b42f008d4eb55a7db9d64999b
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>