Add 1.0-17 to supported stx-openstack app version list
The formal CENGN build uses the --label option when building
the application tarballs, resulting in application versions of:
- 1.0-17-centos-stable-versioned
- 1.0-17-centos-stable-latest

These versions were added to the new application version check
in the following update:
https://review.opendev.org/670754

When the --label is not specified, however, the version does not
include a label. So most developer application build will just
have a version set to "1.0-17" (as a current example).

This update adds this version to the supported version list to
allow for developers using custom builds.

Change-Id: Ifb26e2f391a0195fc53d65b8adf8f87cb87f68e5
Closes-Bug: 1837105
Signed-off-by: Don Penney <don.penney@windriver.com>