Create basic set of unit tests for sysinv ihosts API
Creating a basic set of unit tests for the sysinv ihosts API.
This increases the coverage of the sys/api/controllers/v1/host.py
file from 19% to 47%. It also allows new testcases for the ihosts
API to be easily added.

One minor code change was required to the actual API itself as it
was not functional under python 3.

Change-Id: Ic40f3bc122fa8659fa6813c604e3433aa44ffa58
Story: 2005939
Task: 34292
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>