Initial framework and unit tests for puppet plugins
Introduces base classes and utilities for setting up the system
in specific configurations that are suitable for testing different
different system and host configurations.

The base classes are used by a new set of Puppet plugin unit
tests to improve the test coverage of the puppet plugins.

Additional unit tests will be required to test specific features
of each plugin that is domain specific or based on a specific
deployment configuration.

Also includes a required change for Python 3 compatibility to
permit py35 tox tests to pass.

Change-Id: Id4aa51ce1a8b66987d6f74a8994f989800243b61
Story: 2005939
Task: 35947
Signed-off-by: Matt Peters <matt.peters@windriver.com>