Ceph: update crush map when storage tier is renamed
Renaming storage tiers is supported as long as they are not
in use (no OSDs attached). Tier name is updated correctly
in the inventory database but the crush map is still using
the old name. This causes issues when trying to use the renamed
storage tier.

Add function to rename crush map buckets referenced by the storage
tier being renamed. Note that changes to the crush map are
incremental. Any exception raised while renaming buckets is
in progress causes the crush map to be rolled back.

Change-Id: Ie8f5162e61d291eed29f2f663ec596cd4402e5d9
Closes-Bug: 1836239
Signed-off-by: Daniel Badea <daniel.badea@windriver.com>