Revert "Revert "Changing tiller pod networking settings to improve swact time""
This reverts commit a5c236dc522c050b036e638955c03074a2963996.

It was thought that setting the TCP timeouts for the cluster
network was enough to address the issues with the helm commands
hanging after a controller swact. This is not the case. In
particular, swacting away from the controller with the
tiller-deploy pod seems to cause tcp connection from that pod to
the kube-apiserver to hang. Putting the tiller-deploy pod back on
the host network "fixes" the issue.

Change-Id: I8f37530e1f615afcffcf6cb1d629518436c99cb9
Related-Bug: 1817941
Partial-Bug: 1837055
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>