On all armada applies provide --enable-chart-cleanup
When optional charts are enabled/disabled, manifest overrides are
generated to include/exclude the chart from it's specific chart_group.
By providing --enable-chart-cleanup to the armada apply, armada will
purge chart releases that it's no longer managing.

Prior to this, after disabling a chart and re-applying the application,
the helm release would need to be manually removed.

Change-Id: I34c9bcfd4eb4903e9dc74ec8228146b6cb96e74c
Depends-On: Ic48a2e053d0de7dacfd6a07d817947e11dc8d596
Closes-Bug: #1833738
Signed-off-by: Robert Church <robert.church@windriver.com>