Raise/clear alarms based on Kubernetes app status
This is the final commit for LP1833323.

A major alarm will be raised when an application upload,
apply/update or remove fails. A warning alarm will be raised
when an application apply/update is in progress to discourage
host maintenance (e.g. unlock) as this could be disruptive to
the operation. The alarm is cleared once the operation is
successfully completed.

During an application update, if a failure occurs, automatic
recovery will kick in to revert back to the original app
version. If this is successful, the alarm that was previously
raised at the point of update failure will be cleared.

In the event sysinv conductor is unexpectedly restarted (e.g.
controller swact, failed audit) while an application operation
is in progress, the application status will be reset and an
application alarm raised/cleared accordingly.

This commit also includes the fix for a minor bug introduced
in commit 3f7bbbd94c71c7b8c615704d7edd8ecd62c2f1c4 to add
application-abort command which has no user impact but would
result in a misleading log.

Tests:
  - Verify a warning alarm is generated at the beginning of
    an application apply/update and is cleared at the end.
  - Induce an upload failure, verify a warning alarm is
    generated.
  - Abort during an application-apply, verify a major alarm
    is generated. Retry the apply, verify the alarm is
    cleared at the end of the apply.
  - Abort during an application-update, verify a major alarm
    is generated at the point of failure. Verify that the
    alarm is cleared once automatic recovery has completed.
  - Abort during an application-remove, verify a major alarm
    is generated. Retry the remove, verify the alarm is
    cleared at the end of the apply.

Closes-Bug: 1833323
Depends-On: I992ab7a788cfab8d52d2e6a498519c591148f588
Change-Id: I9bd2cabe318b75b88768ea493992ffd37fb777b0
Signed-off-by: Tee Ngo <Tee.Ngo@windriver.com>