Revert "Fix sysinv handle bmc password cannot be stored case"
This reverts commit 5a9a373297b9a2358c1cb704c2f61075a3ae1905.

Please note that change is suspected to trigger: https://bugs.launchpad.net/starlingx/+bug/1839668

I also encountered host-bulk-add issue in another lab from bug report, and did not encounter it when I reverted this update. This may be coincidental, however, the traceback in the launchpad 1839668 is in this area: Please confirm with retest with host-bulk-add with at least 6 hosts with bmc provisioned. Also, the return code from timeout per 1839668 should be addressed.

Change-Id: I8a6a355d41fd2de2a721cc74a93534012ba7665a