Add an additional check for data interface configuration
If an interface is data interface and the vswitch type is not
accelerated, the data interface is a member of a linux bridge
and then the linux bridge will occupy the interface after host
unlock. if so, it will cause conflict with ovs bridge when ovs
agent add the interface to ovs bridge.
this patch add an additional check to determine if the vswitch
type supports accelerated interfaces. If it doesn`t support
accelerated interfaces, a bridge should not be created.

Change-Id: I0fbae866c364fe7b787aa850db79bca1bf597389
Closes-bug: #1834556
Signed-off-by: marvin <weifei.yu@intel.com>