Configuration out-of-date alarms on storage nodes since fresh install
After fresh install of a Standard configuration with dedicated storage
nodes, the storage node configuration out-of-date alarms are not
cleared. This regression is caused by commit
https://review.opendev.org/#/c/673336/. In check_unlock_storage()
function a call to restore_ceph_config() was replaced by
update_ceph_osd_config() when osdmap is empty. This will generate a
config uuid and invoke config_apply_runtime_manifest before the
.initial_config_complete flag file is created on the storage node.
In this case sysinv-agent won't send response to sysinv-conductor to
clear the configuration out-of-date alarm. The solution is to remove
the code that calls update_ceph_osd_config() in check_unlock_storage()
because ceph config is restored and osds are created automatically
when storage node is unlocked.

Following tests are performed:
  1. Install a storage lab and verify that ceph is running ok without
     storage node configuration out-of-date alarm
  2. Backup and restore a standard storage configuration

Change-Id: If05adc5d30f7ea73cc5e630b4508be0351e4f6e4
Closes-Bug: 1838652
Signed-off-by: Wei Zhou <wei.zhou@windriver.com>
(cherry picked from commit 1ab1aba4b0b90fd9205ebbc14847796636e5725b)