Resolve issue of operation ceph-mon-add blocked
The issue is side effect for patch which resolve ceph-mon-modify issue
with LP: 1827119, since code rework and the case of ceph-mon-add is
not fully considered.

To resolve it, we should make sure cgtsvg_growth_gib == 0 when new ceph
monitor is added, and ceph_mon should not get by uuid with current host
since it is not monitor yet, we should get the first ceph monitor from
monitor list and get the correct value of cgtsvg_growth_gib.

Closes-Bug: 1827080
Change-Id: I18af83864d80111c9b499720e01a1f4302e65d40
Signed-off-by: Chen, Tingjie <tingjie.chen@intel.com>