Allow locking simplex controller
After installing a controller, if it is unlocked before reconfiguring
the required interfaces, the installation cannot be completed.
This commit to allow locking the controller while it is in simplex
state. This commit also reject unlocking controller in a duplex system
if the management and cluster-host interfaces are on the loopback
interface.

After making the active controller on a duplex system return to a
simplex state by locking and removing a standby controller, the address
of the cluster-host interface was not unallocated properly. This commit
corrected this problem.

Closes-Bug: 1830082

Change-Id: I8cb5d1dfd7d4ba73f40aefe3242ca32d91b4e7e8
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>