ceph: enable storage model if storage nodes provisioned
As the system is provisioned by adding new nodes the storage
model can change between controller based or storage based
depending on the type of nodes being added an how the ceph
monitors are configured.

In the case of a storage lab with (for example) 2 controllers,
2 storage nodes and 2 workers configured:
- ceph monitors are configured to run on: controller-0,
  controller-1 and storage-0
- ceph osd's are configured on: storage-0 and storage-1
When storage-0 is locked and deleted then if we look only at
controller and worker nodes there is no difference between this
setup and one with 2 controllers and 2 computes that's being
provisioned.

To fix this issue, look for other provisioned storage nodes.
If any is found then storage model was configured.

Change-Id: Ie23721a24a8b643d4ac26256fe8532806f2672da
Closes-bug: 1840073
Signed-off-by: Daniel Badea <daniel.badea@windriver.com>