Retry puppet runtime apply in case of missing info
There are cases where sysinv-agents gets runtime puppet
apply requests before it gets the host_uuid from the
sysinv-conductor.

This causes the received request to be dropped because the
agent thinks the command is not for itself.

The fix for this is to expand the retry logic of the
config_apply_runtime_manifest method to retry if any
of the needed inventory info is needed.

Tests done:
- install of an SX setup and 10 lock-unlocks
- install of a 2+2 setup and 2 lock-unlocks on each node

Change-Id: I130e73881bd572398072aff7c46b3b7f165ca6b9
Closes-bug: 1841620
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>