Unsuppress and cleanup pep8 whitespace errors
Unsuppressing the following pep8/flake8 errors:
 - E226 missing whitespace around arithmetic operator
 - E402 module level import not at top of file
 - E501 line too long

E402 did not require any code changes.

For line-length E501, the max line length is now set in
tox.ini to 120. This will eventually be reduced to 80,
however such a change should be done in its own task
as it will impact many files and make for a large review
changeset on its own.

Story: 2004515
Task: 36437
Change-Id: Ibe4565f725d2a8fa035eb37261d4f264a4834e21
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>