Update tox.ini for relocated dependencies
This change updates the dependency paths referenced
by tox.ini so that zuul jobs in sysinv can pass.

The STX prefix is no longer needed by tox and zuul
now that the manifest and repo names are the same,
which was fixed as part of the package re-org.

integ is no longer a required project for zuul.
The packages from integ are in utilities instead.

Change-Id: I0ff834e51667e21cbd75262a2c4178fb837f93ad
Story: 2006166
Task: 36488
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>