Minor zuul and tox file cleanup after manifest re-org
The repo re-org has removed 'stx-' from the name of the repos
that are checked out from the manifest, so there is no
need for special prefix handling in tox to accomodate
differences with checking out using "repo" or "git clone".

With the addition of tsconfig to this config repo, it will
generate an intermediate .stestr folder if tox is run
in that subdir.  This has been added to the top level
.gitignore file.

The required-projects for the zuul jobs have been
cleaned up and references to STX-PREFIX are removed.

Change-Id: I8454305867445ad8b0feb9f46d44c7a570e9142a
Story: 2006166
Task: 36490
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>