Elastic: Remove references to stx release prefix
Config should not have to know about the release prefix
of the monitoring application.  A change is pending to
auto configure the initial master nodes (which have this prefix
in the name), and use an environment variable for the
kube-state-metrics host.  This update removes the initial
master nodes config, and references the new kube-state-metrics
host environment variable.

Story: 2005733
Task: 36586
Depends-On: https://review.opendev.org/#/c/681770
Change-Id: I02563563a1100e3be64ace45987b96d759279050
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>