Update default hugepage size in grub
Only a single huge page size is supported now, the default
hugepagesz in the grub config is updated based on the page size
allocated. The 1G huge pages setting was updated only when the
vswitch huge pages were changed, which is now updated when 1G
huge pages are allocated.

Story: 2006295
Task: 36006

Change-Id: Ic8026bdc77131acaafe1cbaa63285ee892c29301
Signed-off-by: Tao Liu <tao.liu@windriver.com>