Disable duplicate address detection when using duplex-direct
In a duplex-direct configuration, the physical mgmt and cluster-host
link on controller-0 will be down until controller-1
comes up. In an IPv6 configuration, this results in dad not completing,
so the addresses stay tentative.

This commit disables duplicate address detection on the mgmt and
cluster-host interface if the system is configured as duplex-direct.
If the interface is a bonded interface, the interface is added to the
bonding masters list before the DAD is disabled.

Closes-Bug: 1836969

Change-Id: I0e169904445db905729fce77e4afa2ba2052598b
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>