Eliminate unnecessary imported modules
In the config project,"/sysinv/sysinv/sysinv/sysinv/db/" directory
has many files imported into the "log" but not used.
At the same time,I built a system of simplex with the latest ISO
of the community. When the system was built,the code was changed.
The system worked well and the other functional modules were not
affected.So I agree we should delete them.

Change-Id: Ibe1f96154d6575d701067e43665b336f3bf5bfec
Partial-Bug: #1844095
Signed-off-by: Hang Li <lihang0166@fiberhome.com>