Allow configuration of isolcpus
Allow a mixture of floating and pinned workloads by allowing the
application cpu's to be split between isolated and non-isolated

CPUs specified as isolated will be added to the isolcpu set in the grub
config. Lowlatency worker nodes will no longer have all application CPUs
in the isolcpu set.

Kubernetes reserves CPUs in a block, not by numa node or range of CPUs.
Due to this we must restrict the Isolated CPUs to a contiguous block
starting on numa 0 (after the platform cores).

Additional work will be required to support isolcpus with vswitch or the
openstack application.

Change-Id: Ia849967e086bb1dadf27020e400a223c4c7cae5f
Story: 2006565
Task: 36664
Signed-off-by: David Sullivan <david.sullivan@windriver.com>