rbd-provisioner storage class exclude 3rd monitor
rbd-provisioner's storage class is referencing all configured
Ceph monitors. When a compute node is deleted and another one is
configured to run the 3rd ceph-mon then the storage class
definition is updated as expected in the overrides but then
platform-integ-apps fails to re-apply because storage class is
immutable (you would need to remove the app first then apply it)

To avoid this issue exclude 3rd monitor from rbd-provisioner's
storage class when generating the overrides.

Change-Id: I546dfc255c5ec362169d23f1804e70b805b2a316
Closes-bug: 1843569
Signed-off-by: Daniel Badea <daniel.badea@windriver.com>