Eliminate unnecessary import modules in the objects directory
In the config project,"/sysinv/sysinv/sysinv/sysinv/objects/" directory
has many files imported into the "log" but not used.
At the same time,I built a system of simplex with the latest ISO
of the community. When the system was built,the code was changed.
The system worked well and the other functional modules were not
affected.So I agree we should delete them.

Change-Id: I35f35ffdf1b9b00a97c259b47420dee3f1e4d939
Partial-Bug: #1844095
Signed-off-by: Hang Li <lihang0166@fiberhome.com>