Fix AIO-DX Controller-1 double reboot after initial unlock
After initial unlocking controller-1, controller-1 rebooted twice.
The second reboot was triggered by the grub audit during the manifest
apply, based on mismatch detection. The grub mismatch was due to the
vswitch 1G memory has not been reflected in the kernel boot arguments
after the initial host-unlock.

This update renames the update_cpu_config method to update_grub_config
and triggers the grub runtime update when the host memory is modified.
This change is also required to ensure the kernel default huge page
size is updated when the host memory is changed.

Closes-Bug: 1841670

Change-Id: I9deb0a443ed86d35b20d5f6779e44204a921dd0d
Signed-off-by: Tao Liu <tao.liu@windriver.com>