Prevent system managed apps from re-applying on host lock
The platform app was re-applying on host-lock of controller-1. This was
blocking additional application actions until the application was
reapplied. With duplex controller installs, typically a controller will
be locked for a maintenance action and then unlocked. There is no need
to reapply the platform application to adjust the replica count for this
scenario.

This change will only evaluate if the application needs to be reapplied
once VIM services are enabled after unlock. This also allows manual
application applies to update the platform apps while one of the duplex
controllers is locked.

Test Scenario #1:
- controller-1 lock:
  - no override eval
  - no app re-apply
- controller-1 unlock:
  - override eval for platform-integ-app after VIM reports services
    enabled
  - no apply needed as the replica count remained 2

Test Scenario #2:
- controller-1 lock:
  - no override eval
  - no app re-apply
- manual apply of platform-integ-app
  - reapplied: 2 provisioners (one pending) -> 1 provisioner
- controller-1 unlock:
  - override eval for platform-integ-app after VIM reports services
    enabled
  - re-apply flag generated: replicas change: 1->2
  - k8s audit reapplies app after 60s

Change-Id: I60fdde2c739f96efb6a5916cc59a1cba43ca5b32
Closes-Bug: #1846056
Signed-off-by: Robert Church <robert.church@windriver.com>